pyproject.toml
[tool.poetry]
name = "django-template-project"
version = "0.1.0"
description = "A template to start hacking with Django"
authors = ["Dan Percic <danpercic86@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.3"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"
Pillow = "^8.2.0"
django-admin-thumbnails = "^0.2.5"
PyYAML = "^5.4.1"
docutils = "^0.16"
django-model-utils = "^4.1.1"
drf-spectacular = "^0.12.0"
django-jazzmin = "^2.4.7"
django-tabular-permissions = "^2.8"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
pylint = "^2.8.2"
black = "^21.5b1"
pycodestyle = "^2.7.0"
django-querycount = "^0.7.0"
django-debug-toolbar = "^3.2.1"
nplusone = "^1.0.0"
django-extra-checks = "^0.8.0"
coverage = { extras = ["toml"], version = "5.5" }
mypy = "^0.812"

[tool.black]
target-version = ['py39']
exclude = '''
/(
    \.git
  | \__pycache__
  | \venv
  | \.venv
  | migrations
)/
'''

[tool.pylint.messages_control]
disable = """
    C0330,
    C0326,
    missing-class-docstring,
    missing-module-docstring,
    too-few-public-methods,
    unused-wildcard-import,
    wildcard-import,
    unsubscriptable-object,
    missing-function-docstring,
    ungrouped-imports,
    no-member,
    too-many-ancestors,
    import-outside-toplevel,
    protected-access,
    invalid-name
"""

[tool.pylint.master]
ignore = """
    migrations
"""

[tool.pylint.format]
max-line-length = "88"

[tool.coverage.run]
omit = [".venv/*", "*/site-packages/*"]

[tool.coverage.report]
fail_under = 100.0

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
